backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
duckdb==0.10.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
umap-learn==0.5.6
networkx==3.3
python-louvain==0.16
anthropic==0.34.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
python-dotenv==1.0.1
pydantic==2.8.2
